move utilities to the test_utils dir
