+++ tests/orchestrator/test_sequential_orchestrator.py
@@ -2,7 +2,7 @@ from azure.durable_functions.models.actions.WhenAnyAction import WhenAnyAction
 from azure.durable_functions.models.actions.WhenAllAction import WhenAllAction
 from azure.durable_functions.models.ReplaySchema import ReplaySchema
 from datetime import datetime, timedelta
-from .orchestrator_test_utils \
+from tests.test_utils.orchestrator_test_utils \
     import assert_orchestration_state_equals, get_orchestration_state_result, assert_valid_schema
 from tests.test_utils.ContextBuilder import ContextBuilder
 from azure.durable_functions.models.OrchestratorState import OrchestratorState
